+++ src/App.js
@@ -1,14 +1,14 @@
 import Header from './components/Header'
-import Card from './components/Card'
+import Collection from './components/Card'
 import { Box } from '@chakra-ui/react'
 import React from 'react'
 
 function App() {
     return (
         <div>
-            <Header />
             <Box maxW="max" ml="auto" mr="auto" mt="24" align="center">
-                <Card header="Initial Caption" body="Initial Text" />
+                <Header />
+                <Collection header="Initial Caption" body="Initial Text" />
             </Box>
         </div>
     )
